Moved actions to single file
